+++ .env
@@ -4,7 +4,7 @@ NEXT_PUBLIC_API_URL=https://beta.api.opensauced.pizza/v2
 NEXT_PUBLIC_EXP_API_URL=https://beta.api.opensauced.pizza/v2
 NEXT_PUBLIC_POSTHOG_ID=phc_Y0xz6nK55MEwWjobJsI2P8rsiomZJ6eZLoXehmMy9tt
 NEXT_PUBLIC_CLOUD_NAME=dgxgziswe
-NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=pk_test_51H1yp3JnZ7CYWRG2KmQMIE8sizoaSaZ8FxlEsc5uQb81gVswDlWu6FErmcukqc6Hso3mhCWgnLsjWphight4t1Mf00jZZjckyZ
+NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=test
 NEXT_PUBLIC_STRIPE_SUB_CANCEL_URL=https://billing.stripe.com/p/login/test_bIY8xy3iy6fpeHu3cc
 NEXT_PUBLIC_OPENGRAPH_URL=https://beta.opengraph.opensauced.pizza/v1
 SENTRY_DSN=https://3f5bb9023ff0407299dd22a6454558f9@o4504872488927232.ingest.sentry.io/4505082236960768
