+++ pages/workspaces/[workspaceId]/repository-insights/[insightId]/dashboard.tsx
@@ -1,3 +1,4 @@
+import dynamic from "next/dynamic";
 import { useEffect, useState } from "react";
 import { GetServerSidePropsContext } from "next";
 import { createPagesServerClient } from "@supabase/auth-helpers-nextjs";
@@ -10,6 +11,9 @@ import useInsightRepositories from "lib/hooks/useInsightRepositories";
 import { WorkspaceLayout } from "components/Workspaces/WorkspaceLayout";
 import HubPageLayout from "layouts/hub-page";
 import Dashboard from "components/organisms/Dashboard/dashboard";
+import { useIsWorkspaceUpgraded } from "lib/hooks/api/useIsWorkspaceUpgraded";
+
+const InsightUpgradeModal = dynamic(() => import("components/Workspaces/InsightUpgradeModal"));
 
 interface InsightPageProps {
   insight: DbUserInsight;
@@ -21,6 +25,10 @@ const HubPage = ({ insight, ogImage, workspaceId }: InsightPageProps) => {
   const { data: insightRepos } = useInsightRepositories(insight.id);
   const repositories = insightRepos.map((repo) => repo.repo_id);
   const [hydrated, setHydrated] = useState(false);
+
+  const { data: isWorkspaceUpgraded } = useIsWorkspaceUpgraded({ workspaceId });
+  const showNudgeBanner = !isWorkspaceUpgraded && repositories.length > 100;
+  const [isInsightUpgradeModalOpen, setIsInsightUpgradeModalOpen] = useState(false);
 
   useEffect(() => {
     setHydrated(true);
@@ -48,9 +56,26 @@ const HubPage = ({ insight, ogImage, workspaceId }: InsightPageProps) => {
         twitterCard="summary_large_image"
       />
       <WorkspaceLayout workspaceId={workspaceId}>
+        {showNudgeBanner && (
+          <button
+            onClick={() => setIsInsightUpgradeModalOpen(true)}
+            className="w-full h-fit py-2 text-center bg-light-orange-10 text-white"
+          >
+            This insight page is past the free limit.{" "}
+            <span className="font-semibold underline">Upgrade to a PRO Workspace!</span>
+          </button>
+        )}
         <HubPageLayout page="dashboard">
           <Dashboard repositories={repositories} />
         </HubPageLayout>
+
+        <InsightUpgradeModal
+          workspaceId={workspaceId}
+          overLimit={repositories.length}
+          isOpen={isInsightUpgradeModalOpen}
+          onClose={() => setIsInsightUpgradeModalOpen(false)}
+          variant="repositories"
+        />
       </WorkspaceLayout>
     </>
   );
